programmes/requirements.txt
TTS
numpy
opencv-python
requests
openai
Pillow
moviepy
google-auth-oauthlib
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
